gpu: nvgpu: fix MISRA 10.3 issues in hal.ltc
Change following ltc hal prototype from:
int (*determine_L2_size_bytes)(struct gk20a *gk20a);
to
u64 (*determine_L2_size_bytes)(struct gk20a *gk20a);

JIRA NVGPU-3422

Reviewed-on: https://git-master.nvidia.com/r/2119996
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/vgpu/ltc/ltc_vgpu.c
@@ -27,7 +27,7 @@
 
 #include "ltc_vgpu.h"
 
-int vgpu_determine_L2_size_bytes(struct gk20a *g)
+u64 vgpu_determine_L2_size_bytes(struct gk20a *g)
 {
 	struct vgpu_priv_data *priv = vgpu_get_priv_data(g);
