+++ drivers/gpu/nvgpu/common/vgpu/ltc/ltc_vgpu.h
@@ -26,7 +26,7 @@
 struct gk20a;
 struct gr_gk20a;
 
-int vgpu_determine_L2_size_bytes(struct gk20a *g);
+u64 vgpu_determine_L2_size_bytes(struct gk20a *g);
 void vgpu_ltc_init_fs_state(struct gk20a *g);
 
 #endif /* NVGPU_LTC_VGPU_H */
